changed type

+++ pie/int64s_pie.go
@@ -30,7 +30,7 @@ func (ss Int64s) Abs() Int64s {
 // as the all() function in Python.
 //
 // If the list is empty then true is always returned.
-func (ss Int64s) All(fn func(value int) bool) bool {
+func (ss Int64s) All(fn func(value int64) bool) bool {
 	for _, value := range ss {
 		if !fn(value) {
 			return false
@@ -44,7 +44,7 @@ func (ss Int64s) All(fn func(value int) bool) bool {
 // as the any() function in Python.
 //
 // If the list is empty then false is always returned.
-func (ss Int64s) Any(fn func(value int) bool) bool {
+func (ss Int64s) Any(fn func(value int64) bool) bool {
 	for _, value := range ss {
 		if fn(value) {
 			return true
@@ -57,7 +57,7 @@ func (ss Int64s) Any(fn func(value int) bool) bool {
 // Append will return a new slice with the elements appended to the end.
 //
 // It is acceptable to provide zero arguments.
-func (ss Int64s) Append(elements ...int) Int64s {
+func (ss Int64s) Append(elements ...int64) Int64s {
 	// Copy ss, to make sure no memory is overlapping between input and
 	// output. See issue #97.
 	result := append(Int64s{}, ss...)
@@ -83,7 +83,7 @@ func (ss Int64s) AreUnique() bool {
 // Average is the average of all of the elements, or zero if there are no
 // elements.
 func (ss Int64s) Average() float64 {
-	if l := int(len(ss)); l > 0 {
+	if l := int64(len(ss)); l > 0 {
 		return float64(ss.Sum()) / float64(l)
 	}
 
@@ -109,7 +109,7 @@ func (ss Int64s) Bottom(n int) (top Int64s) {
 // Contains returns true if the element exists in the slice.
 //
 // When using slices of pointers it will only compare by address, not value.
-func (ss Int64s) Contains(lookingFor int) bool {
+func (ss Int64s) Contains(lookingFor int64) bool {
 	for _, s := range ss {
 		if lookingFor == s {
 			return true
@@ -170,7 +170,7 @@ func (ss Int64s) DropTop(n int) (drop Int64s) {
 
 	// Copy ss, to make sure no memory is overlapping between input and
 	// output. See issue #145.
-	drop = make([]int, len(ss)-n)
+	drop = make([]int64, len(ss)-n)
 	copy(drop, ss[n:])
 
 	return
@@ -178,8 +178,8 @@ func (ss Int64s) DropTop(n int) (drop Int64s) {
 
 // Drop items from the slice while f(item) is true.
 // Afterwards, return every element until the slice is empty. It follows the same logic as the dropwhile() function from itertools in Python.
-func (ss Int64s) DropWhile(f func(s int) bool) (ss2 Int64s) {
-	ss2 = make([]int, len(ss))
+func (ss Int64s) DropWhile(f func(s int64) bool) (ss2 Int64s) {
+	ss2 = make([]int64, len(ss))
 	copy(ss2, ss)
 	for i, value := range ss2 {
 		if !f(value) {
@@ -204,7 +204,7 @@ func (ss Int64s) DropWhile(f func(s int) bool) (ss2 Int64s) {
 //       car.Color = "Red"
 //   })
 //
-func (ss Int64s) Each(fn func(int)) Int64s {
+func (ss Int64s) Each(fn func(int64)) Int64s {
 	for _, s := range ss {
 		fn(s)
 	}
@@ -250,7 +250,7 @@ func (ss Int64s) Extend(slices ...Int64s) (ss2 Int64s) {
 // true from the condition. The returned slice may contain zero elements (nil).
 //
 // FilterNot works in the opposite way of Filter.
-func (ss Int64s) Filter(condition func(int) bool) (ss2 Int64s) {
+func (ss Int64s) Filter(condition func(int64) bool) (ss2 Int64s) {
 	for _, s := range ss {
 		if condition(s) {
 			ss2 = append(ss2, s)
@@ -262,7 +262,7 @@ func (ss Int64s) Filter(condition func(int) bool) (ss2 Int64s) {
 // FilterNot works the same as Filter, with a negated condition. That is, it will
 // return a new slice only containing the elements that returned false from the
 // condition. The returned slice may contain zero elements (nil).
-func (ss Int64s) FilterNot(condition func(int) bool) (ss2 Int64s) {
+func (ss Int64s) FilterNot(condition func(int64) bool) (ss2 Int64s) {
 	for _, s := range ss {
 		if !condition(s) {
 			ss2 = append(ss2, s)
@@ -276,7 +276,7 @@ func (ss Int64s) FilterNot(condition func(int) bool) (ss2 Int64s) {
 // It follows the same logic as the findIndex() function in Javascript.
 //
 // If the list is empty then -1 is always returned.
-func (ss Int64s) FindFirstUsing(fn func(value int) bool) int {
+func (ss Int64s) FindFirstUsing(fn func(value int64) bool) int {
 	for idx, value := range ss {
 		if fn(value) {
 			return idx
@@ -287,13 +287,13 @@ func (ss Int64s) FindFirstUsing(fn func(value int) bool) int {
 }
 
 // First returns the first element, or zero. Also see FirstOr().
-func (ss Int64s) First() int {
+func (ss Int64s) First() int64 {
 	return ss.FirstOr(0)
 }
 
 // FirstOr returns the first element or a default value if there are no
 // elements.
-func (ss Int64s) FirstOr(defaultValue int) int {
+func (ss Int64s) FirstOr(defaultValue int64) int64 {
 	if len(ss) == 0 {
 		return defaultValue
 	}
@@ -321,8 +321,8 @@ func (ss Int64s) Float64s() Float64s {
 
 // Group returns a map of the value with an individual count.
 //
-func (ss Int64s) Group() map[int]int {
-	group := map[int]int{}
+func (ss Int64s) Group() map[int64]int {
+	group := map[int64]int{}
 	for _, n := range ss {
 		group[n]++
 	}
@@ -338,9 +338,9 @@ func (ss Int64s) Intersect(slices ...Int64s) (ss2 Int64s) {
 		return nil
 	}
 
-	var uniqs = make([]map[int]struct{}, len(slices))
+	var uniqs = make([]map[int64]struct{}, len(slices))
 	for i := 0; i < len(slices); i++ {
-		m := make(map[int]struct{})
+		m := make(map[int64]struct{})
 		for _, el := range slices[i] {
 			m[el] = struct{}{}
 		}
@@ -365,7 +365,7 @@ func (ss Int64s) Intersect(slices ...Int64s) (ss2 Int64s) {
 }
 
 // Insert a value at an index
-func (ss Int64s) Insert(index int, values ...int) Int64s {
+func (ss Int64s) Insert(index int, values ...int64) Int64s {
 	if index >= ss.Len() {
 		return Int64s.Extend(ss, Int64s(values))
 	}
@@ -394,7 +394,7 @@ func (ss Int64s) Ints() Ints {
 
 // Join returns a string from joining each of the elements.
 func (ss Int64s) Join(glue string) (s string) {
-	var slice interface{} = []int(ss)
+	var slice interface{} = []int64(ss)
 
 	if y, ok := slice.([]string); ok {
 		// The stdlib is efficient for type []string
@@ -473,12 +473,12 @@ func (ss Int64s) JSONStringIndent(prefix, indent string) string {
 }
 
 // Last returns the last element, or zero. Also see LastOr().
-func (ss Int64s) Last() int {
+func (ss Int64s) Last() int64 {
 	return ss.LastOr(0)
 }
 
 // LastOr returns the last element or a default value if there are no elements.
-func (ss Int64s) LastOr(defaultValue int) int {
+func (ss Int64s) LastOr(defaultValue int64) int64 {
 	if len(ss) == 0 {
 		return defaultValue
 	}
@@ -497,12 +497,12 @@ func (ss Int64s) Len() int {
 // Be careful when using this with slices of pointers. If you modify the input
 // value it will affect the original slice. Be sure to return a new allocated
 // object or deep copy the existing one.
-func (ss Int64s) Map(fn func(int) int) (ss2 Int64s) {
+func (ss Int64s) Map(fn func(int64) int64) (ss2 Int64s) {
 	if ss == nil {
 		return nil
 	}
 
-	ss2 = make([]int, len(ss))
+	ss2 = make([]int64, len(ss))
 	for i, s := range ss {
 		ss2[i] = fn(s)
 	}
@@ -511,7 +511,7 @@ func (ss Int64s) Map(fn func(int) int) (ss2 Int64s) {
 }
 
 // Max is the maximum value, or zero.
-func (ss Int64s) Max() (max int) {
+func (ss Int64s) Max() (max int64) {
 	if len(ss) == 0 {
 		return
 	}
@@ -531,9 +531,9 @@ func (ss Int64s) Max() (max int) {
 //
 // Zero is returned if there are no elements in the slice.
 //
-// If the number of elements is even, then the int mean of the two "median values"
+// If the number of elements is even, then the int64 mean of the two "median values"
 // is returned.
-func (ss Int64s) Median() int {
+func (ss Int64s) Median() int64 {
 	n := len(ss)
 	if n == 0 {
 		return 0
@@ -592,7 +592,7 @@ func (ss Int64s) Median() int {
 }
 
 // Min is the minimum value, or zero.
-func (ss Int64s) Min() (min int) {
+func (ss Int64s) Min() (min int64) {
 	if len(ss) == 0 {
 		return
 	}
@@ -615,7 +615,7 @@ func (ss Int64s) Mode() Int64s {
 	if len(ss) == 0 {
 		return nil
 	}
-	values := make(map[int]int)
+	values := make(map[int64]int)
 	for _, s := range ss {
 		values[s]++
 	}
@@ -646,7 +646,7 @@ func (ss Int64s) Mode() Int64s {
 //   for greeting := greetings.Pop(); greeting != nil; greeting = greetings.Pop() {
 //       fmt.Println(*greeting)
 //   }
-func (ss *Int64s) Pop() (popped *int) {
+func (ss *Int64s) Pop() (popped *int64) {
 
 	if len(*ss) == 0 {
 		return
@@ -658,7 +658,7 @@ func (ss *Int64s) Pop() (popped *int) {
 }
 
 // Product is the product of all of the elements.
-func (ss Int64s) Product() (product int) {
+func (ss Int64s) Product() (product int64) {
 	if len(ss) == 0 {
 		return
 	}
@@ -671,7 +671,7 @@ func (ss Int64s) Product() (product int) {
 }
 
 // Random returns a random element by your rand.Source, or zero
-func (ss Int64s) Random(source rand.Source) int {
+func (ss Int64s) Random(source rand.Source) int64 {
 	n := len(ss)
 
 	// Avoid the extra allocation.
@@ -689,9 +689,9 @@ func (ss Int64s) Random(source rand.Source) int {
 // Reduce continually applies the provided function
 // over the slice. Reducing the elements to a single value.
 //
-// Returns a zero value of int if there are no elements in the slice. It will panic if the reducer is nil and the slice has more than one element (required to invoke reduce).
+// Returns a zero value of int64 if there are no elements in the slice. It will panic if the reducer is nil and the slice has more than one element (required to invoke reduce).
 // Otherwise returns result of applying reducer from left to right.
-func (ss Int64s) Reduce(reducer func(int, int) int) (el int) {
+func (ss Int64s) Reduce(reducer func(int64, int64) int64) (el int64) {
 	if len(ss) == 0 {
 		return
 	}
@@ -714,7 +714,7 @@ func (ss Int64s) Reverse() Int64s {
 		return ss
 	}
 
-	sorted := make([]int, len(ss))
+	sorted := make([]int64, len(ss))
 	for i := 0; i < len(ss); i++ {
 		sorted[i] = ss[len(ss)-i-1]
 	}
@@ -728,7 +728,7 @@ func (ss Int64s) Reverse() Int64s {
 // it locks execution of gorutine
 // it doesn't close channel after work
 // returns sended elements if len(this) != len(old) considered func was canceled
-func (ss Int64s) Send(ctx context.Context, ch chan<- int) Int64s {
+func (ss Int64s) Send(ctx context.Context, ch chan<- int64) Int64s {
 	for i, s := range ss {
 		select {
 		case <-ctx.Done():
@@ -756,8 +756,8 @@ func (ss Int64s) Send(ctx context.Context, ch chan<- int) Int64s {
 // where min is the first param, max is the second, step is the third one, [min, max) with step,
 // others params will be ignored
 func (ss Int64s) Sequence(params ...int) Int64s {
-	var creator = func(i int) int {
-		return int(i)
+	var creator = func(i int) int64 {
+		return int64(i)
 	}
 
 	return ss.SequenceUsing(creator, params...)
@@ -777,7 +777,7 @@ func (ss Int64s) Sequence(params ...int) Int64s {
 // if len(params) > 2 considered that will be returned slice between min and max with step,
 // where min is the first param, max is the second, step is the third one, [min, max) with step,
 // others params will be ignored
-func (ss Int64s) SequenceUsing(creator func(int) int, params ...int) Int64s {
+func (ss Int64s) SequenceUsing(creator func(int) int64, params ...int) Int64s {
 	var seq = func(min, max, step int) (seq Int64s) {
 		lenght := int(util.Round(float64(max-min) / float64(step)))
 		if lenght < 1 {
@@ -805,7 +805,7 @@ func (ss Int64s) SequenceUsing(creator func(int) int, params ...int) Int64s {
 }
 
 // Shift will return two values: the shifted value and the rest slice.
-func (ss Int64s) Shift() (int, Int64s) {
+func (ss Int64s) Shift() (int64, Int64s) {
 	return ss.First(), ss.DropTop(1)
 }
 
@@ -821,7 +821,7 @@ func (ss Int64s) Shuffle(source rand.Source) Int64s {
 	// go 1.10+ provides rnd.Shuffle. However, to support older versions we copy
 	// the algorithm directly from the go source: src/math/rand/rand.go below,
 	// with some adjustments:
-	shuffled := make([]int, n)
+	shuffled := make([]int64, n)
 	copy(shuffled, ss)
 
 	rnd := rand.New(source)
@@ -914,11 +914,11 @@ func (ss Int64s) SubSlice(start int, end int) (subSlice Int64s) {
 		if end <= length {
 			subSlice = ss[start:end]
 		} else {
-			zeroArray := make([]int, end-length)
+			zeroArray := make([]int64, end-length)
 			subSlice = ss[start:length].Append(zeroArray[:]...)
 		}
 	} else {
-		zeroArray := make([]int, end-start)
+		zeroArray := make([]int64, end-start)
 		subSlice = zeroArray[:]
 	}
 
@@ -926,7 +926,7 @@ func (ss Int64s) SubSlice(start int, end int) (subSlice Int64s) {
 }
 
 // Sum is the sum of all of the elements.
-func (ss Int64s) Sum() (sum int) {
+func (ss Int64s) Sum() (sum int64) {
 	for _, s := range ss {
 		sum += s
 	}
@@ -947,7 +947,7 @@ func (ss Int64s) Top(n int) (top Int64s) {
 }
 
 // StringsUsing transforms each element to a string.
-func (ss Int64s) StringsUsing(transform func(int) string) Strings {
+func (ss Int64s) StringsUsing(transform func(int64) string) Strings {
 	l := len(ss)
 
 	// Avoid the allocation.
@@ -980,7 +980,7 @@ func (ss Int64s) Unique() Int64s {
 		return ss
 	}
 
-	values := map[int]struct{}{}
+	values := map[int64]struct{}{}
 
 	for _, value := range ss {
 		values[value] = struct{}{}
@@ -996,7 +996,7 @@ func (ss Int64s) Unique() Int64s {
 
 // Unshift adds one or more elements to the beginning of the slice
 // and returns the new slice.
-func (ss Int64s) Unshift(elements ...int) (unshift Int64s) {
+func (ss Int64s) Unshift(elements ...int64) (unshift Int64s) {
 	unshift = append(Int64s{}, elements...)
 	unshift = append(unshift, ss...)
